Moving iree/base/flags.h to iree/base/internal/.
