+++ iree/hal/vulkan/registration/BUILD
@@ -36,9 +36,9 @@ cc_library(
     ],
     deps = [
         "//iree/base:core_headers",
-        "//iree/base:flags",
         "//iree/base:status",
         "//iree/base:tracing",
+        "//iree/base/internal:flags",
         "//iree/hal:api",
         "//iree/hal/vulkan",
         "@com_google_absl//absl/flags:flag",
