+++ iree/hal/vulkan/registration/driver_module.cc
@@ -17,7 +17,7 @@
 #include <inttypes.h>
 
 #include "absl/flags/flag.h"
-#include "iree/base/flags.h"
+#include "iree/base/internal/flags.h"
 #include "iree/base/status.h"
 #include "iree/base/target_platform.h"
 #include "iree/base/tracing.h"
