+++ iree/tools/BUILD
@@ -30,10 +30,10 @@ cc_binary(
     testonly = True,
     srcs = ["iree-benchmark-module-main.cc"],
     deps = [
-        "//iree/base:flags",
         "//iree/base:status",
         "//iree/base:tracing",
         "//iree/base/internal:file_io",
+        "//iree/base/internal:flags",
         "//iree/hal/drivers",
         "//iree/modules/hal",
         "//iree/tools/utils:vm_util",
@@ -54,10 +54,10 @@ cc_binary(
     deps = [
         "//iree/base:api",
         "//iree/base:core_headers",
-        "//iree/base:flags",
         "//iree/base:status",
         "//iree/base:tracing",
         "//iree/base/internal:file_io",
+        "//iree/base/internal:flags",
         "//iree/hal/drivers",
         "//iree/modules/check:native_module",
         "//iree/modules/hal",
@@ -240,9 +240,9 @@ cc_binary(
         ":init_passes_and_dialects",
         ":init_targets",
         "//iree/base:api",
-        "//iree/base:flags",
         "//iree/base:status",
         "//iree/base:tracing",
+        "//iree/base/internal:flags",
         "//iree/compiler/Dialect/Flow/Transforms",
         "//iree/compiler/Dialect/HAL/Transforms",
         "//iree/compiler/Dialect/IREE/Transforms",
@@ -271,10 +271,10 @@ cc_binary(
     name = "iree-run-module",
     srcs = ["iree-run-module-main.cc"],
     deps = [
-        "//iree/base:flags",
         "//iree/base:status",
         "//iree/base:tracing",
         "//iree/base/internal:file_io",
+        "//iree/base/internal:flags",
         "//iree/hal/drivers",
         "//iree/modules/hal",
         "//iree/tools/utils:vm_util",
